+++ be/src/main/java/codesquad/kr/gyeonggidoidle/issuetracker/domain/auth/controller/response/JwtTokenType.java
@@ -1,0 +1,6 @@
+package codesquad.kr.gyeonggidoidle.issuetracker.domain.auth.controller.response;
+
+public enum JwtTokenType {
+
+    ACCESS, REFRESH
+}
