+++ be/src/main/java/codesquad/kr/gyeonggidoidle/issuetracker/domain/auth/entity/Jwt.java
@@ -1,4 +1,4 @@
-package codesquad.kr.gyeonggidoidle.issuetracker.domain.jwt.entity;
+package codesquad.kr.gyeonggidoidle.issuetracker.domain.auth.entity;
 
 import lombok.Builder;
 import lombok.Getter;
