--- be/src/main/java/codesquad/kr/gyeonggidoidle/issuetracker/domain/jwt/controller/response/JwtTokenType.java
@@ -1,6 +1,0 @@
-package codesquad.kr.gyeonggidoidle.issuetracker.domain.jwt.controller.response;
-
-public enum JwtTokenType {
-
-    ACCESS, REFRESH
-}
